[TTS] move module from model file.
