+++ nemo/collections/tts/models/__init__.py
@@ -14,10 +14,12 @@
 
 from nemo.collections.tts.models.aligner import AlignerModel
 from nemo.collections.tts.models.fastpitch import FastPitchModel
+from nemo.collections.tts.models.fastpitch_ssl import FastPitchModel_SSL
 from nemo.collections.tts.models.hifigan import HifiGanModel
 from nemo.collections.tts.models.mixer_tts import MixerTTSModel
 from nemo.collections.tts.models.radtts import RadTTSModel
 from nemo.collections.tts.models.spectrogram_enhancer import SpectrogramEnhancerModel
+from nemo.collections.tts.models.ssl_tts import SSLDisentangler
 from nemo.collections.tts.models.tacotron2 import Tacotron2Model
 from nemo.collections.tts.models.two_stages import GriffinLimModel, MelPsuedoInverseModel, TwoStagesModel
 from nemo.collections.tts.models.univnet import UnivNetModel
@@ -27,6 +29,8 @@ from nemo.collections.tts.models.waveglow import WaveGlowModel
 __all__ = [
     "AlignerModel",
     "FastPitchModel",
+    "FastPitchModel_SSL",
+    "SSLDisentangler",
     "GriffinLimModel",
     "HifiGanModel",
     "MelPsuedoInverseModel",
